refactor: move consumption type to common
