+++ src/libecalc/common/consumption_type.py
@@ -1,0 +1,6 @@
+from enum import Enum
+
+
+class ConsumptionType(str, Enum):
+    FUEL = "FUEL"
+    ELECTRICITY = "ELECTRICITY"
