+++ src/libecalc/core/consumers/legacy_consumer/component.py
@@ -9,6 +9,7 @@ from numpy.typing import NDArray
 from typing_extensions import assert_never
 
 from libecalc.common.component_type import ComponentType
+from libecalc.common.consumption_type import ConsumptionType
 from libecalc.common.list.list_utils import array_to_list
 from libecalc.common.logger import logger
 from libecalc.common.temporal_model import TemporalExpression, TemporalModel
@@ -43,7 +44,6 @@ from libecalc.core.result.results import (
     GenericComponentResult,
     PumpResult,
 )
-from libecalc.dto.types import ConsumptionType
 from libecalc.expression import Expression
