+++ src/libecalc/presentation/yaml/mappers/component_mapper.py
@@ -5,9 +5,10 @@ from pydantic import TypeAdapter, ValidationError
 
 from libecalc import dto
 from libecalc.common.component_type import ComponentType
+from libecalc.common.consumption_type import ConsumptionType
 from libecalc.common.logger import logger
 from libecalc.common.time_utils import Period, define_time_model_for_period
-from libecalc.dto.types import ConsumerType, ConsumptionType, EnergyModelType
+from libecalc.dto.types import ConsumerType, EnergyModelType
 from libecalc.dto.utils.validators import convert_expression
 from libecalc.expression import Expression
 from libecalc.presentation.yaml.mappers.consumer_function_mapper import (
